+++ src/Table.jsx
@@ -28,9 +28,11 @@ const Table = React.createClass({
     columnsPageSize: PropTypes.number,
     expandIconColumnIndex: PropTypes.number,
     showHeader: PropTypes.bool,
+    title: PropTypes.func,
     footer: PropTypes.func,
     scroll: PropTypes.object,
     rowRef: PropTypes.func,
+    getBodyWrapper: PropTypes.func,
   },
 
   getDefaultProps() {
@@ -64,13 +66,14 @@ const Table = React.createClass({
       rowRef() {
         return null;
       },
+      getBodyWrapper: body => body,
     };
   },
 
   getInitialState() {
     const props = this.props;
     let expandedRowKeys = [];
-    let rows = [ ...props.data ];
+    let rows = [...props.data];
     if (props.defaultExpandAllRows) {
       for (let i = 0; i < rows.length; i++) {
         const row = rows[i];
@@ -92,16 +95,13 @@ const Table = React.createClass({
   },
 
   componentDidMount() {
-    if (this.refs.headTable) {
-      this.refs.headTable.scrollLeft = 0;
-    }
-    if (this.refs.bodyTable) {
-      this.refs.bodyTable.scrollLeft = 0;
-    }
+    this.resetScrollY();
     this.syncFixedTableRowHeight();
     const isAnyColumnsFixed = this.isAnyColumnsFixed();
     if (isAnyColumnsFixed) {
-      this.resizeEvent = addEventListener(window, 'resize', debounce(this.syncFixedTableRowHeight, 150));
+      this.resizeEvent = addEventListener(
+        window, 'resize', debounce(this.syncFixedTableRowHeight, 150)
+      );
     }
   },
 
@@ -110,6 +110,9 @@ const Table = React.createClass({
       this.setState({
         data: nextProps.data,
       });
+      if (!nextProps.data || nextProps.data.length === 0) {
+        this.resetScrollY();
+      }
     }
     if ('expandedRowKeys' in nextProps) {
       this.setState({
@@ -136,9 +139,7 @@ const Table = React.createClass({
 
   onExpandedRowsChange(expandedRowKeys) {
     if (!this.props.expandedRowKeys) {
-      this.setState({
-        expandedRowKeys: expandedRowKeys,
-      });
+      this.setState({ expandedRowKeys });
     }
     this.props.onExpandedRowsChange(expandedRowKeys);
   },
@@ -175,7 +176,7 @@ const Table = React.createClass({
     if (typeof rowKey === 'function') {
       return rowKey(record, index);
     }
-    return record[rowKey] || index;
+    return typeof record[rowKey] !== 'undefined' ? record[rowKey] : index;
   },
 
   getExpandedRows() {
@@ -183,7 +184,7 @@ const Table = React.createClass({
   },
 
   getHeader(columns, fixed) {
-    const { showHeader, expandIconAsCell, prefixCls} = this.props;
+    const { showHeader, expandIconAsCell, prefixCls } = this.props;
     let ths = [];
     if (expandIconAsCell && fixed !== 'right') {
       ths.push({
@@ -212,10 +213,13 @@ const Table = React.createClass({
     const prefixCls = this.props.prefixCls;
     return (
       <tr
-        key={key + '-extra-row'}
-        style={{display: visible ? '' : 'none'}}
-        className={`${prefixCls}-expanded-row ${className}`}>
-        {(this.props.expandIconAsCell && fixed !== 'right') ? <td key="rc-table-expand-icon-placeholder" /> : null}
+        key={`${key}-extra-row`}
+        style={{ display: visible ? '' : 'none' }}
+        className={`${prefixCls}-expanded-row ${className}`}
+      >
+        {(this.props.expandIconAsCell && fixed !== 'right')
+           ? <td key="rc-table-expand-icon-placeholder" />
+           : null}
         <td colSpan={this.props.columns.length}>
           {fixed !== 'right' ? content : '&nbsp;'}
         </td>
@@ -250,7 +254,7 @@ const Table = React.createClass({
       }
       let className = rowClassName(record, i);
       if (this.state.currentHoverKey === key) {
-        className += ' ' + props.prefixCls + '-row-hover';
+        className += ` ${props.prefixCls}-row-hover`;
       }
 
       const onHoverProps = {};
@@ -292,10 +296,14 @@ const Table = React.createClass({
       const subVisible = visible && isRowExpanded;
 
       if (expandedRowContent && isRowExpanded) {
-        rst.push(this.getExpandedRow(key, expandedRowContent, subVisible, expandedRowClassName(record, i), fixed));
+        rst.push(this.getExpandedRow(
+          key, expandedRowContent, subVisible, expandedRowClassName(record, i), fixed
+        ));
       }
       if (childrenColumn) {
-        rst = rst.concat(this.getRowsByData(childrenColumn, subVisible, indent + 1, columns, fixed));
+        rst = rst.concat(this.getRowsByData(
+          childrenColumn, subVisible, indent + 1, columns, fixed
+        ));
       }
     }
     return rst;
@@ -308,7 +316,12 @@ const Table = React.createClass({
   getColGroup(columns, fixed) {
     let cols = [];
     if (this.props.expandIconAsCell && fixed !== 'right') {
-      cols.push(<col className={`${this.props.prefixCls}-expand-icon-col`} key="rc-table-expand-icon-col"></col>);
+      cols.push(
+        <col
+          className={`${this.props.prefixCls}-expand-icon-col`}
+          key="rc-table-expand-icon-col"
+        />
+      );
     }
     cols = cols.concat((columns || this.props.columns).map(c => {
       return <col key={c.key} style={{ width: c.width, minWidth: c.width }} />;
@@ -332,7 +345,7 @@ const Table = React.createClass({
         }
         if (i < pageIndexStart || i > pageIndexEnd) {
           newColumn.className = newColumn.className || '';
-          newColumn.className += ' ' + prefixCls + '-column-hidden';
+          newColumn.className += ` ${prefixCls}-column-hidden`;
         }
         newColumn = this.wrapPageColumn(newColumn, (i === pageIndexStart), (i === pageIndexEnd));
       }
@@ -362,7 +375,7 @@ const Table = React.createClass({
 
   getTable(options = {}) {
     const { columns, fixed } = options;
-    const { prefixCls, scroll = {} } = this.props;
+    const { prefixCls, scroll = {}, getBodyWrapper } = this.props;
     let { useFixedHeader } = this.props;
     const bodyStyle = { ...this.props.bodyStyle };
     const headStyle = {};
@@ -374,7 +387,13 @@ const Table = React.createClass({
     }
 
     if (scroll.y) {
-      bodyStyle.height = bodyStyle.height || scroll.y;
+      // maxHeight will make fixed-Table scrolling not working
+      // so we only set maxHeight to body-Table here
+      if (fixed) {
+        bodyStyle.height = bodyStyle.height || scroll.y;
+      } else {
+        bodyStyle.maxHeight = bodyStyle.maxHeight || scroll.y;
+      }
       bodyStyle.overflowY = bodyStyle.overflowY || 'scroll';
       useFixedHeader = true;
 
@@ -396,15 +415,16 @@ const Table = React.createClass({
           tableStyle.width = scroll.x;
         }
       }
+      const tableBody = hasBody ? getBodyWrapper(
+        <tbody className={`${prefixCls}-tbody`}>
+          {this.getRows(columns, fixed)}
+        </tbody>
+      ) : null;
       return (
         <table className={tableClassName} style={tableStyle}>
           {this.getColGroup(columns, fixed)}
           {hasHead ? this.getHeader(columns, fixed) : null}
-          {hasBody ? (
-            <tbody className={`${prefixCls}-tbody`}>
-              {this.getRows(columns, fixed)}
-            </tbody>
-          ) : null}
+          {tableBody}
         </table>
       );
     };
@@ -418,7 +438,8 @@ const Table = React.createClass({
           style={headStyle}
           onMouseOver={this.detectScrollTarget}
           onTouchStart={this.detectScrollTarget}
-          onScroll={this.handleBodyScroll}>
+          onScroll={this.handleBodyScroll}
+        >
           {renderTable(true, false)}
         </div>
       );
@@ -431,7 +452,8 @@ const Table = React.createClass({
         ref="bodyTable"
         onMouseOver={this.detectScrollTarget}
         onTouchStart={this.detectScrollTarget}
-        onScroll={this.handleBodyScroll}>
+        onScroll={this.handleBodyScroll}
+      >
         {renderTable(!useFixedHeader)}
       </div>
     );
@@ -448,13 +470,15 @@ const Table = React.createClass({
       BodyTable = (
         <div
           className={`${prefixCls}-body-outer`}
-          style={{ ...bodyStyle }}>
+          style={{ ...bodyStyle }}
+        >
           <div
             className={`${prefixCls}-body-inner`}
             ref={refName}
             onMouseOver={this.detectScrollTarget}
             onTouchStart={this.detectScrollTarget}
-            onScroll={this.handleBodyScroll}>
+            onScroll={this.handleBodyScroll}
+          >
             {renderTable(!useFixedHeader)}
           </div>
         </div>
@@ -462,6 +486,15 @@ const Table = React.createClass({
     }
 
     return <span>{headTable}{BodyTable}</span>;
+  },
+
+  getTitle() {
+    const { title, prefixCls } = this.props;
+    return title ? (
+      <div className={`${prefixCls}-title`}>
+        {title(this.state.data)}
+      </div>
+    ) : null;
   },
 
   getFooter() {
@@ -514,6 +547,15 @@ const Table = React.createClass({
     });
   },
 
+  resetScrollY() {
+    if (this.refs.headTable) {
+      this.refs.headTable.scrollLeft = 0;
+    }
+    if (this.refs.bodyTable) {
+      this.refs.bodyTable.scrollLeft = 0;
+    }
+  },
+
   prevColumnsPage() {
     this.goToColumnsPage(this.state.currentColumnsPage - 1);
   },
@@ -538,11 +580,11 @@ const Table = React.createClass({
     const nextHandler = <span className={nextHandlerCls} onClick={this.nextColumnsPage}></span>;
     if (hasPrev) {
       column.title = <span>{prevHandler}{column.title}</span>;
-      column.className = (column.className || '') + ` ${prefixCls}-column-has-prev`;
+      column.className = `${column.className || ''} ${prefixCls}-column-has-prev`;
     }
     if (hasNext) {
       column.title = <span>{column.title}{nextHandler}</span>;
-      column.className = (column.className || '') + ` ${prefixCls}-column-has-next`;
+      column.className = `${column.className || ''} ${prefixCls}-column-has-next`;
     }
     return column;
   },
@@ -584,7 +626,9 @@ const Table = React.createClass({
     if ('isAnyColumnsRightFixedCache' in this) {
       return this.isAnyColumnsRightFixedCache;
     }
-    this.isAnyColumnsRightFixedCache = this.getCurrentColumns().some(column => column.fixed === 'right');
+    this.isAnyColumnsRightFixedCache = this.getCurrentColumns().some(
+      column => column.fixed === 'right'
+    );
     return this.isAnyColumnsRightFixedCache;
   },
 
@@ -605,7 +649,8 @@ const Table = React.createClass({
       if (e.target.scrollLeft === 0) {
         this.setState({ scrollPosition: 'left' });
       } else if (e.target.scrollLeft + 1 >=
-        e.target.children[0].getBoundingClientRect().width - e.target.getBoundingClientRect().width) {
+        e.target.children[0].getBoundingClientRect().width -
+        e.target.getBoundingClientRect().width) {
         this.setState({ scrollPosition: 'right' });
       } else if (this.state.scrollPosition !== 'middle') {
         this.setState({ scrollPosition: 'middle' });
@@ -625,15 +670,9 @@ const Table = React.createClass({
   },
 
   handleRowHover(isHover, key) {
-    if (isHover) {
-      this.setState({
-        currentHoverKey: key,
-      });
-    } else {
-      this.setState({
-        currentHoverKey: null,
-      });
-    }
+    this.setState({
+      currentHoverKey: isHover ? key : null,
+    });
   },
 
   render() {
@@ -642,7 +681,7 @@ const Table = React.createClass({
 
     let className = props.prefixCls;
     if (props.className) {
-      className += ' ' + props.className;
+      className += ` ${props.className}`;
     }
     if (props.columnsPageRange) {
       className += ` ${prefixCls}-columns-paging`;
@@ -656,18 +695,21 @@ const Table = React.createClass({
 
     return (
       <div className={className} style={props.style}>
-        {this.isAnyColumnsLeftFixed() &&
-        <div className={`${prefixCls}-fixed-left`}>
-          {this.getLeftFixedTable()}
-        </div>}
-        <div className={isTableScroll ? `${prefixCls}-scroll` : ''}>
-          {this.getTable()}
-          {this.getFooter()}
+        {this.getTitle()}
+        <div className={`${prefixCls}-content`}>
+          {this.isAnyColumnsLeftFixed() &&
+          <div className={`${prefixCls}-fixed-left`}>
+            {this.getLeftFixedTable()}
+          </div>}
+          <div className={isTableScroll ? `${prefixCls}-scroll` : ''}>
+            {this.getTable()}
+            {this.getFooter()}
+          </div>
+          {this.isAnyColumnsRightFixed() &&
+          <div className={`${prefixCls}-fixed-right`}>
+            {this.getRightFixedTable()}
+          </div>}
         </div>
-        {this.isAnyColumnsRightFixed() &&
-        <div className={`${prefixCls}-fixed-right`}>
-          {this.getRightFixedTable()}
-        </div>}
       </div>
     );
   },